Fix more tests.
