+++ loc/tests/test_admin.py
@@ -28,7 +28,7 @@ def test_loc_actions_shows_pdf_link_when_user_has_letter_request():
     user = UserFactory()
     lr = LetterRequest(user=user)
     lr.save()
-    assert f'/loc/admin/{user.pk}/letter.pdf' in LetterRequestInline.loc_actions(None, lr)
+    assert f'/loc/admin/{user.pk}/finished-letter.pdf' in LetterRequestInline.loc_actions(None, lr)
 
 
 @pytest.mark.django_db
@@ -40,7 +40,7 @@ def test_print_loc_envelopes_works():
 
 
 @pytest.mark.django_db
-def test_letter_request_form_creates_html_content_upon_creation():
+def test_letter_request_form_creates_html_content_upon_creation(allow_lambda_http):
     form1 = save_letter_request_form(form_class=LetterRequestForm)
     lr = form1.instance
     assert 'generated by a staff member' in lr.html_content
@@ -85,7 +85,7 @@ class TestLobIntegrationField:
 
 def create_valid_letter_request():
     user = create_user_with_all_info()
-    return LetterRequestFactory(user=user)
+    return LetterRequestFactory(user=user, html_content='<p>boop</p>')
 
 
 class TestCreateMailConfirmationContext:
