+++ loc/tests/test_schema.py
@@ -4,7 +4,11 @@ from freezegun import freeze_time
 from users.tests.factories import UserFactory
 from onboarding.tests.factories import OnboardingInfoFactory
 from .test_landlord_lookup import mock_lookup_success, enable_fake_landlord_lookup
-from .factories import create_user_with_all_info, LandlordDetailsV2Factory
+from .factories import (
+    create_user_with_all_info,
+    create_user_with_finished_letter,
+    LandlordDetailsV2Factory
+)
 
 
 DEFAULT_ACCESS_DATES_INPUT = {
@@ -224,7 +228,7 @@ def test_landlord_details_are_created_when_user_has_onboarding_info(
 
 
 @pytest.mark.django_db
-def test_letter_request_works(graphql_client, smsoutbox):
+def test_letter_request_works(graphql_client, smsoutbox, allow_lambda_http):
     graphql_client.request.user = create_user_with_all_info()
 
     result = execute_lr_mutation(graphql_client)
@@ -267,7 +271,7 @@ def test_letter_request_is_null_when_user_has_not_yet_requested_letter(graphql_c
 
 
 def test_email_letter_works(db, graphql_client, mailoutbox):
-    graphql_client.request.user = UserFactory.create()
+    graphql_client.request.user = create_user_with_finished_letter()
     result = graphql_client.execute(
         """
         mutation {
